+++ src/xor_addr.rs
@@ -1,4 +1,3 @@
-use anyhow::Result;
 use std::fmt;
 use std::net::{IpAddr, Ipv4Addr};
